cosillas

+++ Mensajeria/src/Persistencia/PersistenciaRutas.java
@@ -55,13 +55,19 @@ public class PersistenciaRutas {
     }
     
     public void guardarRuta(Object x, String data, boolean verificada, String nombreCiudad) throws IOException {
-        File directorio = new File ("Data/Rutas/");
+         File directorio = new File ("Data/Rutas/");
+         System.out.println("LLEGO");
         File[] nombres = directorio.listFiles();
         String nombreFichero;
+        System.out.println("Fitxer que hi han " + nombres.length);
         for(File file:nombres) {
+            System.out.println("entro al for");
             nombreFichero = file.getName();
+            //a data estem tenim en compte el cost, i no l'hem de tenir en compte
             boolean borrar = nombreFichero.startsWith(nombreCiudad + "-" + data);
-            if (borrar) {
+            boolean borrar2 = nombreFichero.endsWith("-NO_verificada-ruta.txt");
+            System.out.println("PASO POR AQUI");
+            if (borrar && borrar2) {
                 file.delete();
             }
         }
@@ -111,7 +117,6 @@ public class PersistenciaRutas {
     void eliminarRutaComp(String inicioRuta, String nombreRuta) {
         File directorio = new File ("Data/Rutas/");
         File[] nombres = directorio.listFiles();
-        ArrayList<String> ficheros = new ArrayList<>();
         String nombreFichero;
         for(File file:nombres) {
             nombreFichero = file.getName();
